+++ src/streetworks/gibraltar/__init__.py
@@ -1,0 +1,5 @@
+"""Gibraltar Street Gazetteer - see :mod:`streetworks.gibraltar.client`."""
+
+from .client import BASE_URL, STREETS_TYPE_NAME, GibraltarStreetsClient
+
+__all__ = ["BASE_URL", "STREETS_TYPE_NAME", "GibraltarStreetsClient"]